Moved cheatcodes to `Cheatnet`
