+++ crates/cheatable-starknet/Cargo.toml
@@ -16,7 +16,15 @@ cairo-felt.workspace = true
 tempfile.workspace = true
 cairo-lang-casm.workspace = true
 cairo-lang-runner.workspace = true
+cairo-lang-utils.workspace = true
+cairo-lang-starknet.workspace = true
 cairo-vm.workspace = true
+thiserror = "1.0.44"
+serde_json = "1.0.104"
+num-traits = "0.2.16"
+starknet = "0.4.0"
+regex = "1.9.1"
+assert_fs = "1.0.13"
 
 [lib]
 name = "cheatable_starknet"
