+++ crates/cheatable-starknet/src/cheatcodes.rs
@@ -1,0 +1,63 @@
+use std::io;
+
+use blockifier::state::errors::StateError;
+use cairo_felt::Felt252;
+use cairo_lang_runner::casm_run::MemBuffer;
+use cairo_vm::vm::errors::hint_errors::HintError;
+use cairo_vm::vm::errors::memory_errors::MemoryError;
+use cairo_vm::vm::errors::vm_errors::VirtualMachineError;
+use starknet_api::StarknetApiError;
+use thiserror::Error;
+
+pub mod declare;
+pub mod deploy;
+pub mod prank;
+pub mod roll;
+pub mod warp;
+
+// All errors that can be thrown from the hint executor have to be added here,
+// to prevent the whole runner from panicking
+#[derive(Error, Debug)]
+pub enum EnhancedHintError {
+    #[error(transparent)]
+    Hint(#[from] HintError),
+    #[error(transparent)]
+    Io(#[from] io::Error),
+    #[error(transparent)]
+    Anyhow(#[from] anyhow::Error),
+    #[error(transparent)]
+    VirtualMachine(#[from] VirtualMachineError),
+    #[error(transparent)]
+    Memory(#[from] MemoryError),
+    #[error(transparent)]
+    State(#[from] StateError),
+    #[error(transparent)]
+    SerdeJson(#[from] serde_json::Error),
+    #[error(transparent)]
+    StarknetApi(#[from] StarknetApiError),
+}
+
+impl From<EnhancedHintError> for HintError {
+    fn from(error: EnhancedHintError) -> Self {
+        match error {
+            EnhancedHintError::Hint(error) => error,
+            error => HintError::CustomHint(error.to_string().into_boxed_str()),
+        }
+    }
+}
+
+fn write_cheatcode_panic(buffer: &mut MemBuffer, panic_data: &[Felt252]) {
+    buffer.write(1).expect("Failed to insert err code");
+    buffer
+        .write(panic_data.len())
+        .expect("Failed to insert panic_data len");
+    buffer
+        .write_data(panic_data.iter())
+        .expect("Failed to insert error in memory");
+}
+
+#[derive(Debug, PartialEq, Clone)]
+pub struct ContractArtifacts {
+    pub sierra: String,
+    pub casm: String,
+}
