--- crates/forge/src/vm_memory.rs
@@ -1,12 +1,0 @@
-use cairo_felt::Felt252;
-use cairo_lang_runner::casm_run::MemBuffer;
-
-pub(crate) fn write_cheatcode_panic(buffer: &mut MemBuffer, panic_data: &[Felt252]) {
-    buffer.write(1).expect("Failed to insert err code");
-    buffer
-        .write(panic_data.len())
-        .expect("Failed to insert panic_data len");
-    buffer
-        .write_data(panic_data.iter())
-        .expect("Failed to insert error in memory");
-}
